Feat:추가 - :data:book BookRepositoryImpl 구현 [S08P22A101-287]

+++ ItsGrey/data/book/src/main/java/com/tntt/book/datasource/RemoteBookDataSource.kt
@@ -5,7 +5,7 @@ import com.tntt.home.model.BookType
 import com.tntt.home.model.SortType
 
 interface RemoteBookDataSource{
-    fun getBookDtos(userId: String, sortType: SortType, startIndex: Int, bookType: BookType): List<BookDto>
+    fun getBookDtos(userId: String, sortType: SortType, startIndex: Long, bookType: BookType): List<BookDto>
     fun createBookDto(userId: String): String
     fun deleteBook(bookIds: List<String>): Boolean
 }
